Reorganise and remove unused file
I think mime.ts makes more sense under client, where it's being
called from. So now server only has the user-facing stuff about
communicating to the server, not the behind-the-scense SOAP client.
